Move Toolbar to components
