+++ pages/mats.js
@@ -4,17 +4,11 @@ import InfiniteScroll from "react-infinite-scroller";
 
 import { matsSearchResults } from "./api/mats";
 import { useSpoilers } from "../hooks/useSpoilers";
-import {
-  baseCharacters,
-  baseUrl,
-  colour,
-  optionToLabel,
-  sortDirectionOptions,
-} from "../data/common";
+import { baseCharacters, baseUrl, colour } from "../data/common";
 
-import Dropdown from "../components/Dropdown";
 import Empty from "../components/Empty";
 import Layout from "../components/Layout";
+import Toolbar from "../components/Toolbar";
 
 const matsPerPage = 5;
 
@@ -23,45 +17,6 @@ const sortOrderOptions = [
   { id: "health", name: "Health" },
   { id: "cards", name: "Cards" },
 ];
-
-function MatsToolbar() {
-  const router = useRouter();
-  const query = router.query;
-
-  function handleSortOrderChange(newOrder) {
-    router.push({
-      pathname: "/mats",
-      query: { ...query, order: newOrder },
-    });
-  }
-
-  function handleSortDirectionChange(newDirection) {
-    router.push({
-      pathname: "/mats",
-      query: { ...query, dir: newDirection },
-    });
-  }
-
-  return (
-    <div className="toolbar">
-      <div className="toolbar-inner">
-        <div className="sort">
-          <Dropdown
-            onChange={handleSortOrderChange}
-            options={sortOrderOptions}
-            value={optionToLabel(query.order, sortOrderOptions)}
-          />
-          <span style={{ margin: "0 8px" }}>:</span>
-          <Dropdown
-            onChange={handleSortDirectionChange}
-            options={sortDirectionOptions}
-            value={optionToLabel(query.dir, sortDirectionOptions)}
-          />
-        </div>
-      </div>
-    </div>
-  );
-}
 
 function Mats({ searchResults }) {
   const { spoilers } = useSpoilers();
@@ -89,7 +44,7 @@ function Mats({ searchResults }) {
 
   return (
     <Layout>
-      <MatsToolbar />
+      <Toolbar pathname="/mats" sortOrderOptions={sortOrderOptions} />
       {matList.length > 0 ? (
         <InfiniteScroll
           className="mat-list"
